requirements.txt
# Requirement files for the ZenComic Downloader
####### Important modules ########
###### Requirements without Version Specifiers ######
requests
beautifulsoup4
pytest
progressbar=1.3
###### END ######
